+++ pubmodules/activities/activityArchiver.js
@@ -42,7 +42,13 @@ class ActivityArchiver {
         activityEvent.on('lead.download.csv', this.save);
 */
       
-        authEvent.on('project_user.invite.pending',  function(event) { 
+      var authProjectUserInvitePendingKey = 'project_user.invite.pending';  //Don't work if job_worker enabled because queue.worker is disabled
+      // if (authEvent.queueEnabled) { //queue not supported.
+      //   authProjectUserInvitePendingKey = 'project_user.invite.pending.queue';
+      // }
+      winston.info('ActivityArchiver authProjectUserInvitePendingKey: ' + authProjectUserInvitePendingKey);
+
+        authEvent.on(authProjectUserInvitePendingKey,  function(event) { 
           setImmediate(() => {
             if (event.skipArchive) {
               return 0;
@@ -57,7 +63,13 @@ class ActivityArchiver {
        
         });
       
-        authEvent.on('project_user.invite',  function(event) { 
+        var authProjectUserInviteKey = 'project_user.invite';   //Don't work if job_worker enabled because queue.worker is disabled
+        // if (authEvent.queueEnabled) { //queue not supported
+        //   authProjectUserInviteKey = 'project_user.invite.queue';
+        // }
+        winston.info('ActivityArchiver authProjectUserInviteKey: ' + authProjectUserInviteKey);
+  
+        authEvent.on(authProjectUserInviteKey,  function(event) { 
           setImmediate(() => {
             if (event.skipArchive) {
               return 0;
@@ -73,9 +85,17 @@ class ActivityArchiver {
           });
         });
       
+      // verified with queue
+        var authProjectUserUpdateKey = 'project_user.update';
+        if (authEvent.queueEnabled) {
+          authProjectUserUpdateKey = 'project_user.update.queue';
+        }
+        winston.info('ActivityArchiver authProjectUserUpdateKey: ' + authProjectUserUpdateKey);
+  
+       authEvent.on(authProjectUserUpdateKey,  function(event) { 
+        setImmediate(() => {   
+          // winston.info('ActivityArchiver authProjectUserUpdateKey: ', event);
       
-       authEvent.on('project_user.update',  function(event) { 
-        setImmediate(() => {         
          /*
          2019-11-20T10:40:52.686991+00:00 app[web.1]: TypeError: Cannot read property '_id' of undefined
 */
@@ -86,20 +106,34 @@ class ActivityArchiver {
           var project_user = undefined;
           if (event.updatedProject_userPopulated.toObject) {
             project_user = event.updatedProject_userPopulated.toObject() 
-          }else {
+          } else {
             project_user = event.updatedProject_userPopulated;
           }
-          var activity = new Activity({actor: {type:"user", id: event.req.user.id, name: event.req.user.fullName }, 
+
+          winston.debug('ActivityArchiver authProjectUserUpdateKey event: ', event);
+
+          if (!event.req.user) {
+            return  winston.debug('ActivityArchiver skipping archive empty user'); //from i think chat21webhook
+          }
+          var activity = new Activity({actor: {type:"user", id: event.req.user._id, name: event.req.user.fullName }, 
                 verb: "PROJECT_USER_UPDATE", actionObj: event.req.body, 
                 target: {type:"project_user", id: event.updatedProject_userPopulated._id.toString(), object: project_user}, 
-                id_project: event.req.projectid });
+                id_project: event.updatedProject_userPopulated.id_project });
           that.save(activity);    
         
       });
         
        });
       
-      authEvent.on('project_user.delete',  function(event) { 
+
+       var authProjectUserDeleteKey = 'project_user.delete';     //Don't work if job_worker enabled because queue.worker is disabled
+      //  if (authEvent.queueEnabled) { //queue not supported
+      //   authProjectUserDeleteKey = 'project_user.delete.queue';
+      //  }
+       winston.info('ActivityArchiver authProjectUserDeleteKey: ' + authProjectUserDeleteKey);
+ 
+
+      authEvent.on(authProjectUserDeleteKey,  function(event) { 
         setImmediate(() => {
           if (event.skipArchive) {
             return 0;
@@ -115,105 +149,159 @@ class ActivityArchiver {
       });
       
 
-        authEvent.on('user.signin',  function(event) { 
-            winston.debug('ActivityArchiver user.login');
-          setImmediate(() => {
+      // var authUserSignineKey = 'user.signin';
+      // // if (authEvent.queueEnabled) { //queue not supported
+      // //   authUserSignineKey = 'user.signin.queue';
+      // // }
+      // winston.info('ActivityArchiver authUserSignineKey: ' + authUserSignineKey);
+
+
+      //   authEvent.on(authUserSignineKey,  function(event) { 
+      //       winston.info('ActivityArchiver user.login');
+      //     setImmediate(() => {
            
             
-            if (event.skipArchive) {
-              return 0;
-            }
+      //       if (event.skipArchive) {
+      //         return 0;
+      //       }
 
-              var activity = new Activity({actor: {type:"user", id: event.user._id, name: event.user.fullName }, 
-                      verb: "USER_SIGNIN", actionObj: event.req.body, 
-                      target: {type:"user", id:event.user._id.toString(), object: null }, 
-                      id_project: '*' });
-                that.save(activity);   
+      //         var activity = new Activity({actor: {type:"user", id: event.user._id, name: event.user.fullName }, 
+      //                 verb: "USER_SIGNIN", actionObj: event.req.body, //insecure it store password
+      //                 target: {type:"user", id:event.user._id.toString(), object: null }, 
+      //                 id_project: '*' }); /// * project
+      //           that.save(activity);   
             
-          });   
-        });
-        authEvent.on('user.login.error',  function(event) { 
-          setImmediate(() => {
-            
-            if (event.skipArchive) {
-              return 0;
-            }
+      //     });   
+      //   });
 
-              var activity = new Activity({actor: {type:"user"}, 
-              verb: "USER_SIGNIN_ERROR", actionObj: event.req.body, 
-              target: {type:"user", id:null, object: null }, 
-              id_project: '*' });
-              that.save(activity);     
-            
-          });
-        });
-      
-       authEvent.on('user.requestresetpassword',  function(event) {   
-        setImmediate(() => {   
-         
-          if (event.skipArchive) {
-            return 0;
-          }
 
-            var activity = new Activity({actor: {type:"user", id: event.updatedUser._id, name: event.updatedUser.fullName }, 
-              verb: "USER_REQUEST_RESETPASSWORD", actionObj: event.req.body, 
-              target: {type:"user", id:event.updatedUser._id.toString(), object: null }, 
-              id_project: '*' });
-            that.save(activity);       
+        // var authUserSigninErrorKey = 'user.login.error';
+        // // if (authEvent.queueEnabled) {  //queue not supported
+        // //   authUserSigninErrorKey = 'user.login.error.queue';
+        // // }
+        // winston.info('ActivityArchiver authUserSigninErrorKey: ' + authUserSigninErrorKey);
+
         
-        });
-       });
+        // authEvent.on(authUserSigninErrorKey,  function(event) { 
+        //   setImmediate(() => {
+            
+        //     if (event.skipArchive) {
+        //       return 0;
+        //     }
+
+        //       var activity = new Activity({actor: {type:"user"}, 
+        //       verb: "USER_SIGNIN_ERROR", actionObj: event.req.body,     //insecure it store password
+        //       target: {type:"user", id:null, object: null }, 
+        //       id_project: '*' });     /// * project
+        //       that.save(activity);     
+            
+        //   });
+        // });
       
-      authEvent.on('user.resetpassword',  function(event) {     
-        setImmediate(() => {     
+
+      //   var authUserResetPasswordKey = 'user.requestresetpassword';
+      //   // if (authEvent.queueEnabled) {  //queue not supported
+      //   //   authUserResetPasswordKey = 'user.requestresetpassword.queue';
+      //   // }
+      //   winston.info('ActivityArchiver authUserResetPasswordKey: ' + authUserResetPasswordKey);
+
+      //  authEvent.on(authUserResetPasswordKey,  function(event) {   
+      //   setImmediate(() => {   
          
-          if (event.skipArchive) {
-            return 0;
-          }
+      //     if (event.skipArchive) {
+      //       return 0;
+      //     }
 
-          var activity = new Activity({actor: {type:"user", id: event.saveUser._id, name: event.saveUser.fullName }, 
-            verb: "USER_RESETPASSWORD", actionObj: null, //req.body otherwise print password  
-            target: {type:"user", id:event.saveUser._id.toString(), object: null }, 
-            id_project: '*' });
-          that.save(activity);     
+      //       var activity = new Activity({actor: {type:"user", id: event.updatedUser._id, name: event.updatedUser.fullName }, 
+      //         verb: "USER_REQUEST_RESETPASSWORD", actionObj: event.req.body, 
+      //         target: {type:"user", id:event.updatedUser._id.toString(), object: null }, 
+      //         id_project: '*' });        /// * project
+      //       that.save(activity);       
+        
+      //   });
+      //  });
+      
+
+      //  var authUserResetPassword2Key = 'user.resetpassword';
+      // //  if (authEvent.queueEnabled) {  //queue not supported
+      // //   authUserResetPassword2Key = 'user.resetpassword.queue';
+      // //  }
+      //  winston.info('ActivityArchiver authUserResetPassword2Key: ' + authUserResetPassword2Key);
+
+
+      // authEvent.on(authUserResetPassword2Key,  function(event) {     
+      //   setImmediate(() => {     
+         
+      //     if (event.skipArchive) {
+      //       return 0;
+      //     }
+
+      //     var activity = new Activity({actor: {type:"user", id: event.saveUser._id, name: event.saveUser.fullName }, 
+      //       verb: "USER_RESETPASSWORD", actionObj: null, //req.body otherwise print password  
+      //       target: {type:"user", id:event.saveUser._id.toString(), object: null }, 
+      //       id_project: '*' });     /// * project
+      //     that.save(activity);     
           
-        });
-      });
+      //   });
+      // });
       
       
-       authEvent.on('user.signup',  function(event) { 
-        setImmediate(() => {      
-          if (event.skipArchive) {
-            return 0;
-          }
+      // var authUserSignupKey = 'user.signup';
+      // //  if (authEvent.queueEnabled) {  //queue not supported
+      // //   authUserSignupKey = 'user.signup.queue';
+      // //  }
+      //  winston.info('ActivityArchiver authUserSignupKey: ' + authUserSignupKey);
 
-            var activity = new Activity({actor: {type:"user", id: event.savedUser._id, name: event.savedUser.fullName }, 
-                  verb: "USER_SIGNUP", actionObj: event.req.body, 
-                  target: {type:"user", id: event.savedUser._id.toString(), object: null }, 
-                  id_project: '*' });
-              that.save(activity);      
+
+      //  authEvent.on(authUserSignupKey,  function(event) { 
+      //   setImmediate(() => {      
+      //     if (event.skipArchive) {
+      //       return 0;
+      //     }
+
+      //       var activity = new Activity({actor: {type:"user", id: event.savedUser._id, name: event.savedUser.fullName }, 
+      //             verb: "USER_SIGNUP", actionObj: event.req.body, 
+      //             target: {type:"user", id: event.savedUser._id.toString(), object: null }, 
+      //             id_project: '*' });     /// * project
+      //         that.save(activity);      
            
-        });
-       });
+      //   });
+      //  });
       
-      authEvent.on('user.signup.error',  function(event) { 
-        setImmediate(() => {     
-          if (event.skipArchive) {
-            return 0;
-          }
+
+      //  var authUserSignupErrorKey = 'user.signup.error';
+      // //  if (authEvent.queueEnabled) {   //queue not supported
+      // //   authUserSignupErrorKey = 'user.signup.error.queue';
+      // //  }
+      //  winston.info('ActivityArchiver authUserSignupErrorKey: ' + authUserSignupErrorKey);
 
 
-          var activity = new Activity({actor: {type:"user"}, 
-              verb: "USER_SIGNUP_ERROR", actionObj: event.req.body, 
-              target: {type:"user", id:null, object: null }, 
-              id_project: '*' });
-            that.save(activity);    
+      // authEvent.on(authUserSignupErrorKey,  function(event) { 
+      //   setImmediate(() => {     
+      //     if (event.skipArchive) {
+      //       return 0;
+      //     }
+
+
+      //     var activity = new Activity({actor: {type:"user"}, 
+      //         verb: "USER_SIGNUP_ERROR", actionObj: event.req.body, 
+      //         target: {type:"user", id:null, object: null }, 
+      //         id_project: '*' });     /// * project
+      //       that.save(activity);    
           
-        });
-       });
+      //   });
+      //  });
       
-        requestEvent.on('request.create',  function(request) {   
-          setImmediate(() => {           
+
+       var requestCreateKey = 'request.create';
+       if (requestEvent.queueEnabled) {
+         requestCreateKey = 'request.create.queue';
+       }
+       winston.info('ActivityArchiver requestCreateKey: ' + requestCreateKey);
+
+        requestEvent.on(requestCreateKey,  function(request) {   
+          setImmediate(() => {        
+            winston.debug('ActivityArchiver requestCreate triggered');   
           // problema requester_id
 
           // Error saving activity  {"_id":"5e06189c6e226d358896d733","actor":{"_id":"5e06189c6e226d358896d734","type":"user","id":null},"verb":"REQUEST_CREATE","actionObj":{"status":200,"participants":["5e06189c6e226d358896d728"],"messages_count":0,"tags":[],"_id":"5e06189c6e226d358896d72e","request_id":"request_id-closeRequest","first_text":"first_text","department":{"routing":"assigned","default":true,"status":1,"_id":"5e06189c6e226d358896d72b","name":"Default Department","id_project":"5e06189c6e226d358896d729","createdBy":"5e06189c6e226d358896d728","createdAt":"2019-12-27T14:43:40.327Z","updatedAt":"2019-12-27T14:43:40.327Z","__v":0},"agents":[{"_id":"5e06189c6e226d358896d72a","id_project":"5e06189c6e226d358896d729","id_user":"5e06189c6e226d358896d728","role":"owner","user_available":true,"createdBy":"5e06189c6e226d358896d728","createdAt":"2019-12-27T14:43:40.324Z","updatedAt":"2019-12-27T14:43:40.324Z","__v":0}],"id_project":"5e06189c6e226d358896d729","createdBy":"requester_id1","channel":{"name":"chat21"},"createdAt":"2019-12-27T14:43:40.586Z","updatedAt":"2019-12-27T14:43:40.586Z","__v":0},"target":{"type":"request","id":"5e06189c6e226d358896d72e","object":{"status":200,"participants":["5e06189c6e226d358896d728"],"messages_count":0,"tags":[],"_id":"5e06189c6e226d358896d72e","request_id":"request_id-closeRequest","first_text":"first_text","department":{"routing":"assigned","default":true,"status":1,"_id":"5e06189c6e226d358896d72b","name":"Default Department","id_project":"5e06189c6e226d358896d729","createdBy":"5e06189c6e226d358896d728","createdAt":"2019-12-27T14:43:40.327Z","updatedAt":"2019-12-27T14:43:40.327Z","__v":0},"agents":[{"_id":"5e06189c6e226d358896d72a","id_project":"5e06189c6e226d358896d729","id_user":"5e06189c6e226d358896d728","role":"owner","user_available":true,"createdBy":"5e06189c6e226d358896d728","createdAt":"2019-12-27T14:43:40.324Z","updatedAt":"2019-12-27T14:43:40.324Z","__v":0}],"id_project":"5e06189c6e226d358896d729","createdBy":"requester_id1","channel":{"name":"chat21"},"createdAt":"2019-12-27T14:43:40.586Z","updatedAt":"2019-12-27T14:43:40.586Z","__v":0}},"id_project":"5e06189c6e226d358896d729"}
@@ -248,7 +336,17 @@ class ActivityArchiver {
         });
 
 
-        requestEvent.on('request.update.preflight',  function(request) {   
+        // verified with queue
+        var requestUpdatePreflightKey = 'request.update.preflight';
+        if (requestEvent.queueEnabled) {
+          requestUpdatePreflightKey = 'request.update.preflight.queue';
+        }
+        winston.info('ActivityArchiver requestUpdatePreflightKey: ' + requestUpdatePreflightKey);
+
+        
+        requestEvent.on(requestUpdatePreflightKey,  function(request) {
+          winston.debug('ActivityArchiver request.update.preflight: ');
+   
           setImmediate(() => {           
        
               try {
@@ -271,7 +369,16 @@ class ActivityArchiver {
         });
 
 
-        requestEvent.on('request.close',  function(request) {   
+
+        // verified with queue
+        var requestCloseKey = 'request.close';
+        if (requestEvent.queueEnabled) {
+          requestCloseKey = 'request.close.queue';
+        }
+        winston.info('ActivityArchiver requestCloseKey: ' + requestCloseKey);
+
+
+        requestEvent.on(requestCloseKey,  function(request) {   
           setImmediate(() => {           
        
               try {